requirements.txt
streamlit
python-dotenv
langchain
langchain-community
langchain-groq
langsmith
langgraph
wikipedia
# Persistence
aiosqlite
langgraph-checkpoint-sqlite
langgraph-checkpoint-postgres
psycopg-binary
psycopg_pool
# MCP
mcp
